Add Coyright headers to the files
Rename file and folder names to be consistent with other language
transaction processors.
Also add copyright headers to all the files uploaded.
