+++ goprocessor/Dockerfile
@@ -1,3 +1,18 @@
+# Copyright 2018 Intel Corporation
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# -----------------------------------------------------------------------------
+
 FROM ubuntu:xenial
 
 RUN echo "deb [arch=amd64] http://repo.sawtooth.me/ubuntu/ci xenial universe" >> /etc/apt/sources.list \
@@ -43,4 +58,4 @@ WORKDIR /project/simplewallet/goprocessor
 COPY . ./
 RUN bash -c "./build.sh"
 RUN rm -rf src/ *.sh Dockerfile
-CMD bash -C "simple_wallet"
+CMD bash -C "simplewallet"
