+++ goprocessor/build.sh
@@ -1,3 +1,21 @@
 #!/usr/bin/env bash
 
-go build src/sawtooth_siimplewallet/simple_wallet.go
+#------------------------------------------------------------------------------
+# Copyright 2018 Intel Corporation
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#------------------------------------------------------------------------------
+
+# Produces a binary file simplewallet
+# We may use optional -o to change file name
+go build src/sawtooth_simplewallet/simplewallet.go
