--- goprocessor/src/sawtooth_siimplewallet/utils/helper.go
@@ -1,14 +1,0 @@
-package utils
-
-import (
-	"crypto/sha512"
-	"encoding/hex"
-	"strings"
-)
-
-func Hexdigest(str string) string {
-	hash := sha512.New()
-	hash.Write([]byte(str))
-	hashBytes := hash.Sum(nil)
-	return strings.ToLower(hex.EncodeToString(hashBytes))
-}
