+++ goprocessor/src/sawtooth_simplewallet/constants/globals.go
@@ -1,0 +1,35 @@
+/**
+ * Copyright 2018 Intel Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * ------------------------------------------------------------------------------
+ */
+
+package constants
+
+const (
+	// Related to transaction family
+	TransactionFamilyName    string = "simplewallet"
+	TransactionFamilyVersion string = "1.0"
+	// Related to length of addresses
+	TransactionFamilyNamespaceAddressLength int = 6
+	TransactionUserAddressLength            int = 64
+	// Constants used in CLI
+	TransactionTransfer string = "transfer"
+	TransactionDeposit  string = "deposit"
+	TransactionWithdraw string = "withdraw"
+	// Transaction indices
+	TransactionOperationIndex int = 0
+	TransactionAmountIndex    int = 1
+	TransactionToIndex        int = 2
+)
