+++ goprocessor/src/sawtooth_simplewallet/handlers/simplewallet_handler.go
@@ -1,3 +1,20 @@
+/**
+ * Copyright 2018 Intel Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * ------------------------------------------------------------------------------
+ */
+
 package handlers
 
 import (
@@ -5,8 +22,8 @@ import (
 	"github.com/hyperledger/sawtooth-sdk-go/logging"
 	"github.com/hyperledger/sawtooth-sdk-go/processor"
 	"github.com/hyperledger/sawtooth-sdk-go/protobuf/processor_pb2"
-	"sawtooth_siimplewallet/constants"
-	"sawtooth_siimplewallet/utils"
+	"sawtooth_simplewallet/constants"
+	"sawtooth_simplewallet/utils"
 	"strconv"
 	"strings"
 )
