+++ goprocessor/src/sawtooth_simplewallet/simplewallet.go
@@ -1,3 +1,20 @@
+/**
+ * Copyright 2018 Intel Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * ------------------------------------------------------------------------------
+ */
+
 package main
 
 import (
@@ -6,7 +23,7 @@ import (
 	processor2 "github.com/hyperledger/sawtooth-sdk-go/processor"
 	"github.com/jessevdk/go-flags"
 	"os"
-	handler2 "sawtooth_siimplewallet/handlers"
+	handler2 "sawtooth_simplewallet/handlers"
 	"syscall"
 )
