+++ goprocessor/src/sawtooth_simplewallet/utils/helper.go
@@ -1,0 +1,31 @@
+/**
+ * Copyright 2018 Intel Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * ------------------------------------------------------------------------------
+ */
+
+package utils
+
+import (
+	"crypto/sha512"
+	"encoding/hex"
+	"strings"
+)
+
+func Hexdigest(str string) string {
+	hash := sha512.New()
+	hash.Write([]byte(str))
+	hashBytes := hash.Sum(nil)
+	return strings.ToLower(hex.EncodeToString(hashBytes))
+}
