+++ simplewallet-build-tp-go.yaml
@@ -1,3 +1,18 @@
+# Copyright 2018 Intel Corporation
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# -----------------------------------------------------------------------------
+
 version: '2.1'
 
 services:
@@ -16,7 +31,7 @@ services:
       - 'no_proxy=rest-api,validator,${no_proxy}'
     depends_on:
       - validator
-    command: "simple_wallet -C tcp://validator:4004"
+    command: "simplewallet -C tcp://validator:4004"
     stop_signal: SIGKILL
 
   simplewallet-client:
